language with activity

+++ tiago_gpt4/script/text_to_speech_gpt4.py
@@ -80,5 +80,6 @@ class TTSFunction:
                 self.stream.stop_stream()
                 self.stream.close()
 
-        # Terminate PyAudio
-        p.terminate()
+            # Terminate PyAudio
+            if p:
+                p.terminate()
